Fix UI tests to work with the simplified UI system
