+++ Assets/Tests/PlayMode/Gameplay.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c71f11877e9556040b8fc6e981f47505
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
