+++ Assets/Tests/PlayMode/Gameplay/BattleTests.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 using UnityEngine.TestTools;
 using UnityEngine.SceneManagement;
 
-namespace Tests
+namespace Tests.Gameplay
 {
     public class BattleTests
     {
