+++ Assets/Tests/PlayMode/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e5f29c8c4b81bd54ba565059617f9d92
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
