+++ Assets/Tests/PlayMode/UI/DialogTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 23483bc70844bfb4c95c8e8b32ff96a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
