+++ Assets/Tests/PlayMode/UI/WalletUITests.cs
@@ -1,0 +1,26 @@
+using System.Collections;
+using UnityEngine.TestTools;
+using NUnit.Framework;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+namespace Tests.UI
+{
+    /// <summary>
+    /// TODO: Tests for the WalletUI
+    /// </summary>
+    public class WalletUITests
+    {
+        [SetUp]
+        public void Setup()
+        {
+            SceneManager.LoadScene("MainScene");
+        }
+
+        [UnityTest]
+        public IEnumerator CanCreateUpgradeUIMenu()
+        {
+            yield return null;
+        }
+    }
+}
