+++ Assets/Tests/PlayMode/UI/WalletUITests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e632593c31a23254bb45833fc41fafc2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
